Flush the headers to the backend as soon as they are received
No need to wait and backend can process them as we read body potentially
slightly increasing performance. Also, if this is a new connection and backend
has aggressive first-byte timeout (such as Varnish in default configuration),
this will prevent the backend from time-ing out when a client has a long delay
between headers and body (as is customary for busy Internet Exporer <= 8).