Remove Grace period option, wait for all worker threads to finish
When worker is required to shut down (by SIGHUP/SIGINT), it waits for all busy
threads to finish. Number of running threads is monitored every 10 seconds.
During the shutdown, control socket is temporary renamed (pid is appended to
the socket name), so that it is still accessible and correct socket is removed
on exit.