copy_chunks: Signal an error if a terminating chunk is missing
This means that the client disconnected abruptly, without completing the
request. If we proceed waiting for reply from backend without ever completing
the request we'll just wait forever.

Note that this has some security implications as it provides a way for the
client to hit and run, wasting a backend connection while disconnecting his one
(which leads to denial of service).